# docs/admin-api/admin_swagger.yaml
basePath: /api/v1
definitions:
  api.PageData:
    properties:
      list: {}
      pageNum:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  api.Response:
    properties:
      code:
        type: integer
      data: {}
      detail:
        type: string
      msg:
        type: string
      requestId:
        type: string
    type: object
  dto.Base64ImageUploadReq:
    properties:
      image:
        type: string
    type: object
...
  dto.ImageUrlUploadReq:
    properties:
      url:
        type: string
    type: object
...
  dto.SysConfigAddReq:
    properties:
      ValueType:
        type: string
      description:
        type: string
      name:
        type: string
      value:
        type: string
    type: object
  dto.SysConfigDeleteReq:
    properties:
      id:
        example: "0"
        type: string
    type: object
  dto.SysConfigLogPageQueryReq:
    properties:
      configId:
        example: "0"
        type: string
      pageNum:
        type: integer
      pageSize:
        type: integer
    type: object
  dto.SysConfigPageQueryReq:
    properties:
      idOrder:
        type: string
      keyword:
        type: string
      pageNum:
        type: integer
      pageSize:
        type: integer
    type: object
  dto.SysConfigUpdateReq:
    properties:
      description:
        type: string
      id:
        example: "0"
        type: string
      value:
        type: string
    type: object
...
  dto.UploadFileResp:
    properties:
      url:
        type: string
    type: object
...
  models.SysConfig:
    properties:
      createdAt:
        type: string
      createdBy:
        example: "0"
        type: string
      description:
        type: string
      id:
        example: "0"
        type: string
      name:
        type: string
      updatedAt:
        type: string
      updatedBy:
        example: "0"
        type: string
      value:
        type: string
      valueType:
        type: string
    type: object
  models.SysConfigLog:
    properties:
      configId:
        type: integer
      createdAt:
        type: string
      createdBy:
        example: "0"
        type: string
      id:
        example: "0"
        type: string
      newValue:
        type: string
      oldValue:
        type: string
    type: object
...
host: 127.0.0.1:10000
info:
  contact:
    email: xxx@qq.com
    name: xxx
    url: http://xxx.com
  description: xxx管理后台API文档
  title: Admin API
  version: 1.0.0
paths:
  /file/download_and_store:
    post:
      consumes:
      - application/json
      description: 下载网络图片并转存
      parameters:
      - description: 网络图片URL
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.ImageUrlUploadReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UploadFileResp'
              type: object
      summary: 下载网络图片并转存
      tags:
      - 文件上传
  /file/upload_base64_image:
    post:
      consumes:
      - application/json
      description: 上传Base64格式图片
      parameters:
      - description: base64格式图片
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.Base64ImageUploadReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UploadFileResp'
              type: object
      summary: 上传Base64格式图片
      tags:
      - 文件上传
  /file/upload_blob_image:
    post:
      consumes:
      - application/json
      description: 上传blob格式图片
      parameters:
      - description: blob格式图片
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.Base64ImageUploadReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UploadFileResp'
              type: object
      summary: 上传blob格式图片
      tags:
      - 文件上传
  /file/upload_image:
    post:
      consumes:
      - multipart/form-data
      description: 上传图片
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UploadFileResp'
              type: object
      summary: 上传图片
      tags:
      - 文件上传
  /sys/config/add:
    post:
      consumes:
      - application/json
      description: 添加配置
      parameters:
      - description: 配置信息
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysConfigAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 添加配置
      tags:
      - 配置管理
  /sys/config/delete:
    post:
      consumes:
      - application/json
      description: 根据id删除配置
      parameters:
      - description: 要删除的配置ID
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysConfigDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 删除配置
      tags:
      - 配置管理
  /sys/config/page_query:
    post:
      consumes:
      - application/json
      description: 分页查询配置
      parameters:
      - description: 配置筛选条件
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysConfigPageQueryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/api.PageData'
                  - properties:
                      List:
                        $ref: '#/definitions/models.SysConfig'
                    type: object
              type: object
      security:
      - RequireLogin: []
      summary: 分页查询配置
      tags:
      - 配置管理
  /sys/config/page_query_log:
    post:
      consumes:
      - application/json
      description: 分页查询配置日志
      parameters:
      - description: 配置日志筛选条件
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysConfigLogPageQueryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/api.PageData'
                  - properties:
                      List:
                        $ref: '#/definitions/models.SysConfigLog'
                    type: object
              type: object
      security:
      - RequireLogin: []
      summary: 分页查询配置日志
      tags:
      - 配置管理
  /sys/config/update:
    post:
      consumes:
      - application/json
      description: 根据id修改配置信息
      parameters:
      - description: 配置信息
        in: body
        name: param
        schema:
          $ref: '#/definitions/dto.SysConfigUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - RequireLogin: []
      summary: 修改配置
      tags:
      - 配置管理
...
securityDefinitions:
  RequireLogin:
    in: header
    name: token
    type: apiKey
swagger: "2.0"